+++ components/FormContainer/FormContainer.css
@@ -1,0 +1,5 @@
+.form_container {
+  display: block;
+  margin: 2.9rem 0 0 0.25rem;
+  height: fit-content;
+}
