+++ components/MainForm/MainForm.css
@@ -1,0 +1,46 @@
+.mainForm {
+  display: block;
+  margin: 2.9rem 0 0 0.25rem;
+  height: fit-content;
+}
+
+.form_buttons {
+  position: absolute;
+  border: 1px solid red;
+  top: 55%;
+}
+.stop_button--passive {
+  display: inline-block;
+  background-color: #a1acb0;
+  border: 2px solid #a1acb0;
+  color: #799970;
+}
+
+.stop_button--active {
+  border: 2px solid var(--red);
+  background-color: var(--primary);
+  color: var(--darkblue);
+}
+
+.save_button {
+  display: inline-block;
+  background-color: #a1acb0;
+  border: 2px solid #a1acb0;
+  color: #799970;
+}
+
+.save_button--active {
+  display: inline-block;
+  background-color: var(--secondary-light);
+  border: 2px solid var(--secondary-light);
+  color: black;
+}
+
+.label_standard {
+  margin-left: 0.25rem;
+}
+
+.input_text {
+  border-radius: 4px;
+  margin: 0.25rem 0.5rem;
+}
