+++ components/MiniClock/MiniClock.js
@@ -1,4 +1,6 @@
 export default function MiniClock(props) {
+  console.log(props);
+  console.log(timingProps);
   const miniClock = document.createElement("section");
   miniClock.classList.add("miniClock");
   miniClock.setAttribute("id", "miniClock");
