+++ styles.css
@@ -7,9 +7,10 @@
 @import "./components/Header/Header.css";
 @import "./components/ListContainer/ListContainer.css";
 @import "./components/ListEntry/ListEntry.css";
-@import "./components/MainTiming/MainTiming.css";
+@import "./components/MainForm/MainForm.css";
 @import "./components/Menu/Menu.css";
 @import "./components/MiniClock/MiniClock.css";
 @import "./components/Modal/Modal.css";
+@import "./components/Timer/Timer.css";
 @import "./components/TimeRecords/TimeRecords.css";
 @import "./components/What/What.css";
